(12) WIP: create message handlers, refactoring
